+++ PDC/PDC_mainCode/PDC_SPI.h
@@ -1,2 +1,2 @@
-unsigned int readSPI(int deviceSelect, byte registerSelect, int numBytes);
-void writeSPI(int deviceSelect, byte registerSelect, int data);
+uint32_t readSPI(uint8_t deviceSelect, uint8_t registerSelect, uint8_t numBytes);
+void writeSPI(uint8_t deviceSelect, uint8_t registerSelect, uint8_t data);
